server/requirements.txt
fastapi>=0.89.0
pydantic>=2.0.3
sqlalchemy==1.4.49
starlette>=0.23.0

loguru~=0.7.0
uvicorn~=0.23.1
passlib~=1.7.4
bcrypt~=4.0.1
python-jose~=3.3.0
python-multipart~=0.0.6
phonenumbers~=8.13.16
psycopg2-binary~=2.9.6
python-dotenv~=1.0.0
pydantic_extra_types~=2.0.0
fastapi_jwt_auth~=0.5.0
fastapi-mail~=1.4.1
websockets~=11.0.3
starlette-admin~=0.11.1
itsdangerous~=2.1.2
Jinja2~=3.1.2
celery
redis
requests
yookassa